feat: move model_config to into prompts and rename conversession module and delete module model_config
