+++ src/rsazure_openai_toolkit/prompts/__init__.py
@@ -1,5 +1,16 @@
-from .model import PromptMetadata, PromptData
+from .model import (
+    PromptMetadata,
+    PromptData,
+    ModelConfig,
+    get_model_config
+)
+
 from .manager import PromptManager
 
-
-__all__ = ["PromptMetadata", "PromptData", "PromptManager"]
+__all__ = [
+    "PromptMetadata",
+    "PromptData",
+    "ModelConfig",
+    "get_model_config",
+    "PromptManager"
+]
